Added support for query string syntax array/object fields in multipart
