+++ lib/http.js
@@ -1,5 +1,6 @@
 var http = require('http');
 var escape = require('querystring').escape;
+var parse = require('./parse');
 
 // TODO: Benchmark zlib performance against node-compress.
 var zlib = require('zlib');
@@ -91,47 +92,6 @@ if (!requestProto.hasOwnProperty('cookies')) {
       return cookies;
     }
   });
-}
-
-function parse(string) {
-  var data;
-  if (typeof string == 'string') {
-    if (string[0] == '{') {
-      try {
-        data = JSON.parse(string);
-      }
-      catch (e) {
-        this.logger.error(e.stack);
-        this.logger.log(buffer);
-      }
-    }
-    else {
-      data = {};
-      string.split('&').forEach(function (pair) {
-        pair = pair.split('=');
-        var key = decodeURIComponent(pair[0]);
-        var value = decodeURIComponent(pair[1]);
-        var isObjectOrArray = false;
-        var subKey = null;
-        key = key.replace(/\[([^\]]*)\]$/, function (match, inner) {
-          isObjectOrArray = true;
-          subKey = inner;
-          return '';
-        });
-        if (key) {
-          if (isObjectOrArray) {
-            data[key] = data[key] || ((subKey && isNaN(subKey)) ? {} : []);
-            subKey = subKey || data[key].length;
-            data[key][subKey] = value;
-          }
-          else {
-            data[key] = value;
-          }
-        }
-      });
-    }
-  }
-  return data || {};
 }
 
 if (!requestProto.hasOwnProperty('query')) {
