+++ lib/multipart.js
@@ -1,3 +1,5 @@
+var parse = require('./parse');
+
 function isBr(n) {
   return n == 13 || n == 10;
 }
@@ -147,7 +149,7 @@ module.exports = function (request, response, maxBytes) {
     }
     else {
       data += d;
-      request.multipart[field.name] = data;
+      parse.applyPair(request.multipart, field.name, data);
     }
   }
